+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.1
+* Added a way to remove active bars in options menu.
+* Fixed yet another module icon nil error.
+
 ## 2.0.0
 * Legion update (no major changes).
 * Fixed icon update bug.